refactor: move check_auto_compact into compact module
